MNGECLIPSE-554: Make project import extensible

+++ org.maven.ide.eclipse/src/org/maven/ide/eclipse/internal/project/MavenMarkerManager.java
@@ -173,7 +173,9 @@ public class MavenMarkerManager {
   }
 
   void deleteMarkers(IFile pom) throws CoreException {
-    pom.deleteMarkers(MavenPlugin.MARKER_ID, true, IResource.DEPTH_INFINITE);
+    if (pom != null && pom.exists()) {
+      pom.deleteMarkers(MavenPlugin.MARKER_ID, true, IResource.DEPTH_INFINITE);
+    }
   }
 
   private void addMissingArtifactMarkers(IFile pomFile, MavenProject mavenProject) {
